requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
cluster==1.4.1.post3
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.11.0
easyocr==1.7.1
editdistance==0.6.2
efficientnet==1.0.0
essential-generators==1.0
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.42.1
fsspec==2023.10.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-vision==3.4.5
googleapis-common-protos==1.61.0
gradio==4.7.1
gradio_client==0.7.0
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.14.0
h5py==3.10.0
htr-pipeline==1.2.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.6
imageio==2.33.0
imgaug==0.4.0
importlib-resources==6.1.1
imutils==0.5.4
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
Keras-Applications==1.0.8
keras-ocr==0.9.3
kiwisolver==1.4.5
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
mediapipe==0.10.5
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.0
onnxruntime==1.16.3
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
orjson==3.9.10
packaging==23.1
pandas==2.1.3
path==16.7.1
Pillow==10.0.1
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
pyreadline3==3.4.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.1
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
semantic-version==2.10.0
shapely==2.0.2
shellingham==1.5.4
six==1.16.0
sklearn==0.0.post11
sniffio==1.3.0
sounddevice==0.4.6
starlette==0.27.0
sympy==1.12
tesseract==0.1.3
threadpoolctl==3.2.0
tifffile==2023.9.26
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
validators==0.22.0
websockets==11.0.3
